requirements.txt
celery==5.3.4
Django==4.2.7
django-kombu==0.9.4
python-dateutil==2.8.2
setuptools==59.6.0
six==1.16.0
tzdata==2023.3
wcwidth==0.2.9
djangorestframework==3.14.0
Markdown==3.4.3
django-filter==23.1


